pyproject.toml
[project]
name = "codeagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.38.0",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
]
